finsihed bot(kinda)
the bot is finished i have the finalised code running on the dcu server
i just forgot to pull it from that. the only change on that was it just
calls the shell script directly and it uses process builder rather then
just runtime.exec(); as this can actually cause a problem with string
argumnets for the shell script not being read correctly.  will pull that
when i get into college and commit it